Moving persistence and config under infrastructure package
