+++ src/main/kotlin/com/paucls/runbookDDD/infrastructure/config/RedisConfig.kt
@@ -1,4 +1,4 @@
-package com.paucls.runbookDDD.config
+package com.paucls.runbookDDD.infrastructure.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
